🏗️ Abstracted pow2 into utils
